+++ CHANGES.md
@@ -1,2 +1,5 @@
 - Added smooth transitions between pulls
-- Fix switching sublevels manually in guide mode
+- Added option to fade out window when mouse isn't over it
+- Speed up route estimation by adding a min-heap for the path queue as well as length and weight limits
+- Improved route estimation accuracy by switching weights to a rolling average
+- Fixed switching sublevels manually in guide mode